Create UpdatedAPI
This API code is fully functional and vetted to be more than just the Outline.  It's intent is to give full file system control in a known local enviornment to the API file itself.

All files and subdirectories it manages is in it's local "uploads" folder which it creates upon launching if it does not already inherently exist.